Renamed rulset collapse component
