+++ client/src/components/layout/ProfileSidebar.js
@@ -10,7 +10,7 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import SidebarImage from './SidebarImage';
 import smashDefaultImage from '../../img/smashultimate1.jpg';
-import RuleSets from './RuleSets';
+import RuleSets from './RuleSetCollapse';
 import UpdateDetailsModal from '../auth/UpdateDetailsModal';
